fix: Renamed conflicting package name
