+++ .gitignore
@@ -132,5 +132,5 @@ ansible/site.retry
 .direnv/
 
 # gRPC
-
 *pb2*
+*.proto
