+++ weni/org_grpc/__init__.py
@@ -1,3 +1,3 @@
-from weni.org_grpc import grpc
+from weni.org_grpc import grpc_gen
 
 default_app_config = "weni.org_grpc.apps.OrgGrpcConfig"
