+++ weni/org_grpc/serializers.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import User
 from rest_framework import serializers
 
 from django_grpc_framework import proto_serializers
-from weni.org_grpc.grpc import org_pb2
+from weni.org_grpc.grpc_gen import org_pb2
 
 from temba.orgs.models import Org
