+++ weni/org_grpc/urls.py
@@ -1,4 +1,4 @@
-from weni.org_grpc.grpc import org_pb2_grpc
+from weni.org_grpc.grpc_gen import org_pb2_grpc
 
 from weni.org_grpc.services import OrgService
